# server/minichat-user/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8081","description":"Generated server url"}],"paths":{"/user/{id}":{"get":{"tags":["用户资料"],"summary":"查询用户资料","operationId":"getUserInfo","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultUserInfoResponse"}}}}}},"put":{"tags":["用户资料"],"summary":"修改用户资料","operationId":"updateUser","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/group/{groupId}":{"put":{"tags":["群组管理"],"summary":"修改群信息（群主操作）","operationId":"updateGroup","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"groupId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateGroup"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}},"delete":{"tags":["群组管理"],"summary":"解散群（群主操作）","operationId":"dismissGroup","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"groupId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/user/avatar":{"post":{"tags":["用户资料"],"summary":"上传头像（OSS）","operationId":"uploadAvatar","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"required":["file"],"type":"object","properties":{"file":{"type":"string","format":"binary"}}}}}},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/group":{"post":{"tags":["群组管理"],"summary":"创建群","operationId":"createGroup","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGroup"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultLong"}}}}}}},"/group/{groupId}/quit":{"post":{"tags":["群组管理"],"summary":"退出群","operationId":"quitGroup","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"groupId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/group/{groupId}/kick/{memberId}":{"post":{"tags":["群组管理"],"summary":"踢出群成员（群主操作）","operationId":"kickMember","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"groupId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"memberId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/friend/{friendId}/block":{"post":{"tags":["好友管理"],"summary":"拉黑/取消拉黑","operationId":"toggleBlock","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"friendId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/friend/request":{"post":{"tags":["好友管理"],"summary":"发送好友请求","operationId":"sendRequest","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendFriendRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/friend/request/{id}/reject":{"post":{"tags":["好友管理"],"summary":"拒绝好友请求","operationId":"rejectRequest","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/friend/request/{id}/accept":{"post":{"tags":["好友管理"],"summary":"同意好友请求","operationId":"acceptRequest","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/auth/sms/code":{"post":{"tags":["用户认证"],"summary":"发送短信验证码","operationId":"sendSmsCode","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SendSmsCodeRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/auth/register":{"post":{"tags":["用户认证"],"summary":"注册","operationId":"register","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/auth/logout":{"post":{"tags":["用户认证"],"summary":"退出登录","operationId":"logout","parameters":[{"name":"Authorization","in":"header","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}},"/auth/login":{"post":{"tags":["用户认证"],"summary":"密码登录","operationId":"login","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginRequest"}}},"required":true},"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultString"}}}}}}},"/internal/user/{id}":{"get":{"tags":["internal-user-controller"],"operationId":"getUserById","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InternalUserResponse"}}}}}}},"/group/{groupId}/members":{"get":{"tags":["群组管理"],"summary":"群成员列表","operationId":"getMembers","parameters":[{"name":"groupId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListGroupMemberVO"}}}}}}},"/group/list":{"get":{"tags":["群组管理"],"summary":"我的群列表","operationId":"getMyGroups","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListGroupVO"}}}}}}},"/friend/search":{"get":{"tags":["好友管理"],"summary":"全局搜索用户（ES）","operationId":"searchUsers","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"keyword","in":"query","required":true,"schema":{"type":"string"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListUserInfoResponse"}}}}}}},"/friend/requests":{"get":{"tags":["好友管理"],"summary":"查看待处理的好友请求","operationId":"getRequests","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListFriendRequest"}}}}}}},"/friend/list":{"get":{"tags":["好友管理"],"summary":"好友列表","operationId":"getFriendList","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultListFriendVO"}}}}}}},"/friend/{friendId}":{"delete":{"tags":["好友管理"],"summary":"删除好友","operationId":"deleteFriend","parameters":[{"name":"userId","in":"header","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"friendId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"400":{"description":"Bad Request","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"500":{"description":"Internal Server Error","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}},"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/ResultVoid"}}}}}}}},"components":{"schemas":{"ResultVoid":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"object"}}},"UpdateUserRequest":{"type":"object","properties":{"nickname":{"maxLength":50,"minLength":0,"type":"string"},"avatar":{"type":"string"},"signature":{"maxLength":255,"minLength":0,"type":"string"},"gender":{"type":"integer","format":"int32"}}},"UpdateGroup":{"type":"object","properties":{"name":{"maxLength":50,"minLength":0,"type":"string"},"notice":{"maxLength":500,"minLength":0,"type":"string"}}},"ResultString":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"string"}}},"CreateGroup":{"required":["memberIds","name"],"type":"object","properties":{"name":{"maxLength":50,"minLength":0,"type":"string"},"memberIds":{"type":"array","items":{"type":"integer","format":"int64"}}}},"ResultLong":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"integer","format":"int64"}}},"SendFriendRequest":{"required":["toId"],"type":"object","properties":{"toId":{"type":"integer","format":"int64"},"remark":{"type":"string"}}},"SendSmsCodeRequest":{"type":"object","properties":{"phone":{"pattern":"^1[3-9]\\d{9}$","type":"string"}}},"RegisterRequest":{"required":["password","smsCode"],"type":"object","properties":{"phone":{"pattern":"^1[3-9]\\d{9}$","type":"string"},"smsCode":{"type":"string"},"password":{"maxLength":20,"minLength":6,"type":"string"}}},"LoginRequest":{"required":["password"],"type":"object","properties":{"phone":{"pattern":"^1[3-9]\\d{9}$","type":"string"},"password":{"type":"string"}}},"ResultUserInfoResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"$ref":"#/components/schemas/UserInfoResponse"}}},"UserInfoResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"phone":{"type":"string"},"nickname":{"type":"string"},"avatar":{"type":"string"},"signature":{"type":"string"},"gender":{"type":"integer","format":"int32"},"createTime":{"type":"string","format":"date-time"}}},"InternalUserResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"phone":{"type":"string"},"nickname":{"type":"string"},"avatar":{"type":"string"},"signature":{"type":"string"},"gender":{"type":"integer","format":"int32"},"status":{"type":"integer","format":"int32"},"lastLoginTime":{"type":"string","format":"date-time"},"lastLoginIp":{"type":"string"},"createTime":{"type":"string","format":"date-time"},"updateTime":{"type":"string","format":"date-time"}}},"GroupMemberVO":{"type":"object","properties":{"userId":{"type":"integer","format":"int64"},"nickname":{"type":"string"},"avatar":{"type":"string"},"role":{"type":"integer","format":"int32"},"alias":{"type":"string"}}},"ResultListGroupMemberVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/GroupMemberVO"}}}},"GroupVO":{"type":"object","properties":{"groupId":{"type":"integer","format":"int64"},"name":{"type":"string"},"ownerId":{"type":"integer","format":"int64"},"ownerNickname":{"type":"string"},"notice":{"type":"string"},"memberCount":{"type":"integer","format":"int32"},"createTime":{"type":"string","format":"date-time"}}},"ResultListGroupVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/GroupVO"}}}},"ResultListUserInfoResponse":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/UserInfoResponse"}}}},"FriendRequest":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"fromId":{"type":"integer","format":"int64"},"toId":{"type":"integer","format":"int64"},"remark":{"type":"string"},"status":{"type":"integer","format":"int32"},"createTime":{"type":"string","format":"date-time"},"updateTime":{"type":"string","format":"date-time"}}},"ResultListFriendRequest":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/FriendRequest"}}}},"FriendVO":{"type":"object","properties":{"friendId":{"type":"integer","format":"int64"},"nickname":{"type":"string"},"avatar":{"type":"string"},"remark":{"type":"string"},"createTime":{"type":"string","format":"date-time"}}},"ResultListFriendVO":{"type":"object","properties":{"code":{"type":"integer","format":"int32"},"msg":{"type":"string"},"data":{"type":"array","items":{"$ref":"#/components/schemas/FriendVO"}}}}}}}